+++ src/templates/expanded-endpoint-template.js
@@ -9,14 +9,6 @@ import '../components/api-response.js';
 
 /* eslint-disable indent */
 export function expandedEndpointBodyTemplate(path, tagName = '') {
-  const acceptContentTypes = new Set();
-  for (const respStatus in path.responses) {
-    for (const acceptContentType in path.responses[respStatus]?.content) {
-      acceptContentTypes.add(acceptContentType.trim());
-    }
-  }
-  const accept = [...acceptContentTypes].join(', ');
-
   // Filter API Keys that are non-empty and are applicable to the the path
   const nonEmptyApiKeys = this.resolvedSpec.securitySchemes.filter((v) => (v.finalKeyValue && path.security && path.security.some((ps) => ps[v.apiKeyId]))) || [];
 
@@ -60,7 +52,6 @@ export function expandedEndpointBodyTemplate(path, tagName = '') {
           fill-defaults = "${!this.hideDefaults}"
           display-nulls="${!!this.includeNulls}"
           enable-console = "${!this.hideExecution}"
-          accept = "${accept}"
           render-style="${this.renderStyle}" 
           schema-style = "${this.displaySchemaAsTable ? 'table' : 'tree'}"
           active-schema-tab = "${this.defaultSchemaTab}"